fix UI table

+++ src/pages/Customer.jsx
@@ -22,16 +22,20 @@ const Customer = () => {
 
   // Function to delete a user
   const deleteUser = async (userId) => {
+    if (!window.confirm("Có chắc muốn xóa người dùng này")) {
+      return;
+    }
+
     try {
       const res = await axios.delete(
-        `${process.env.REACT_APP_API_URL}/admin/manage/:${userId}`
+        `${process.env.REACT_APP_API_URL}/admin/manage/${userId}`
       );
       console.log(res.data);
-      setUsers(users.filter((user) => user.id !== userId));
-      alert("Có chắc muốn xóa người dùng này");
+      setUsers((prevUsers) => prevUsers.filter((user) => user.id !== userId));
       toast.success("User deleted successfully!");
     } catch (error) {
-      console.error(error);
+      console.error("Failed to delete user:", error);
+      toast.error("Failed to delete user.");
     }
   };
 
@@ -42,36 +46,38 @@ const Customer = () => {
   return (
     <div className="dashboard-container">
       <h1 className="dashboard-heading">Manage Users</h1>
-      <table className="user-table">
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Email</th>
-            <th>Role</th>
-            <th>Action</th>
-          </tr>
-        </thead>
-        <tbody>
-          {users.map(
-            (user) =>
-              user.role === "user" && (
-                <tr key={user.id} className="user-item">
-                  <td>{user.name}</td>
-                  <td>{user.email}</td>
-                  <td>{user.role}</td>
-                  <td>
-                    <button
-                      className="delete-btn"
-                      onClick={() => deleteUser(user.id)}
-                    >
-                      Delete
-                    </button>
-                  </td>
-                </tr>
-              )
-          )}
-        </tbody>
-      </table>
+      <div className="table-wrapper">
+        <table className="user-table">
+          <thead>
+            <tr>
+              <th>Name</th>
+              <th>Email</th>
+              <th>Role</th>
+              <th>Action</th>
+            </tr>
+          </thead>
+          <tbody>
+            {users.map(
+              (user) =>
+                user.role === "user" && (
+                  <tr key={user.id} className="user-item">
+                    <td>{user.name}</td>
+                    <td>{user.email}</td>
+                    <td>{user.role}</td>
+                    <td>
+                      <button
+                        className="delete-btn"
+                        onClick={() => deleteUser(user.id)}
+                      >
+                        Delete
+                      </button>
+                    </td>
+                  </tr>
+                )
+            )}
+          </tbody>
+        </table>
+      </div>
     </div>
   );
 };
